This includes all the work by Pat McCormick <pat@lanl.gov> to add support for
AMDGPU.s This also includes an overhaul of the kitsune/experiments directory
and everything in it.

There is still some work to be done - for instance, getting rid of the legacy
pass manager for everything except backend code generation, but this should
be in a functional state for now.

+++ llvm/lib/Transforms/Tapir/GPUABI.cpp
@@ -373,7 +373,7 @@ void LLVMLoop::processOutlinedLoopCall(TapirLoopInfo &TL, TaskOutlineInfo &TOI,
   bcw.writeStrtab(); 
 
   char* heapbuf = new char[mbuf.size()];
-  for(int i=0; i<mbuf.size(); i++) { heapbuf[i] = mbuf[i]; } 
+  for(size_t i=0; i<mbuf.size(); i++) { heapbuf[i] = mbuf[i]; } 
 
   std::string strbuf(mbuf.data(), mbuf.size()); 
   std::ofstream out("compile_time.bc");
